requirements.txt
absl-py==2.1.0
altair==5.4.1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.2.10
fonttools==4.54.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.13.0
idna==3.8
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.8.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
MouseInfo==0.1.3
namex==0.0.8
narwhals==1.6.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.28.0
pure_eval==0.2.3
pyarrow==17.0.0
PyAutoGUI==0.9.54
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.1
pywin32==308
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
setuptools==75.8.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
stack-data==0.6.3
streamlit==1.40.1
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarg==0.1.9
